trying to fix post form modal to render correct post's body
