+++ frontend/src/components/Newsfeed/index.js
@@ -3,8 +3,9 @@ import Posts from '../Posts/index';
 import './Newsfeed.css'
 
 const Newsfeed = () => {
+
     const currentUser = JSON.parse(sessionStorage.getItem('currentUser'));
-    
+
     return (
         <>
             <div className="left-side-bar-container">
