+++ frontend/src/components/Posts/Posts.css
@@ -223,7 +223,7 @@
     position: fixed;
     height: 100vh;
     width: 200vw;
-    background-color: rgba(255, 255, 255, 0.8);
+    background-color: rgba(255, 255, 255, 0.2);
     display: flex;
     justify-content: center;
     align-items: center;
